requirements.txt
antlr4-python3-runtime==4.9.3
antlr-denter
snakes